Added scan for changes
added state comparison between stop and starting the service. Running into limitations of the file system watcher, the buffers mean some events are missed and seems unreliable. The best solution will be using this a notification and then we can manually scan the folder for changes. We will need to catalogue folders and files on system start and scan for changes when notified.